requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
multidict==6.0.5
PyJWT==2.9.0
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.2
yarl==1.9.7
